[RENAME]: AddFriendTVC -> FriendAddTVC, 구조 변경 취소
